Rollout v176 is now 'optional', in that it checks to see if the Resource Types it adds already exist in the DB

+++ lib/rollout/version/Flex_Rollout_Version_000176.php
@@ -4,7 +4,8 @@
  * Version 176 of database update.
  * This version: -
  *	
- *	1:	Add the SFTP and iSeek Data File Resource Types
+ *	1:	Add the SFTP Resource Type
+ *	2:	Add the iSeek Data File Resource Type
  *
  */
 
@@ -17,19 +18,63 @@ class Flex_Rollout_Version_000176 extends Flex_Rollout_Version
 		$dbAdmin = Data_Source::get(FLEX_DATABASE_CONNECTION_ADMIN);
 		$dbAdmin->setFetchMode(MDB2_FETCHMODE_ASSOC);
 		
-		//	1:	Add the SFTP and iSeek Data File Resource Types
-		$strSQL = "	INSERT INTO	resource_type
-						(name, description, const_name, resource_type_nature)
-					VALUES
-						('SFTP File Server'			, 'SFTP File Server'		, 'RESOURCE_TYPE_FILE_RESOURCE_SFTP'			, (SELECT id FROM resource_type_nature WHERE const_name = 'RESOURCE_TYPE_NATURE_FILE_REPOSITORY' LIMIT 1)),
-						('iSeek Data Usage File'	, 'iSeek Data Usage File'	, 'RESOURCE_TYPE_FILE_IMPORT_CDR_ISEEK_DATA'	, (SELECT id FROM resource_type_nature WHERE const_name = 'RESOURCE_TYPE_NATURE_IMPORT_FILE' LIMIT 1));";
-		$result = $dbAdmin->query($strSQL);
+		//	1:	Add the SFTP Resource Type
+		$strSQL	= "	SELECT	id
+					FROM	resource_type
+					WHERE	const_name = 'RESOURCE_TYPE_FILE_RESOURCE_SFTP'
+					LIMIT	1;";
+		$result	= $dbAdmin->query($strSQL);
 		if (PEAR::isError($result))
 		{
-			throw new Exception(__CLASS__ . ' Failed to add the SFTP and iSeek Data File Resource Types. ' . $result->getMessage() . " (DB Error: " . $result->getUserInfo() . ")");
+			throw new Exception(__CLASS__ . ' Failed to check whether SFTP Resource Type already exists . ' . $result->getMessage() . " (DB Error: " . $result->getUserInfo() . ")");
 		}
-		$this->rollbackSQL[] =	"	DELETE FROM	resource_type
-									WHERE		const_name IN ('RESOURCE_TYPE_FILE_RESOURCE_SFTP', 'RESOURCE_TYPE_FILE_IMPORT_CDR_ISEEK_DATA');";
+		elseif (!$result->fetchRow())
+		{
+			// DEBUG
+			throw new Exception("SFTP already exists!");
+			// DEBUG
+			
+			$strSQL = "	INSERT INTO	resource_type
+							(name, description, const_name, resource_type_nature)
+						VALUES
+							('SFTP File Server'			, 'SFTP File Server'		, 'RESOURCE_TYPE_FILE_RESOURCE_SFTP'			, (SELECT id FROM resource_type_nature WHERE const_name = 'RESOURCE_TYPE_NATURE_FILE_REPOSITORY' LIMIT 1));";
+			$result = $dbAdmin->query($strSQL);
+			if (PEAR::isError($result))
+			{
+				throw new Exception(__CLASS__ . ' Failed to add the SFTP Resource Type. ' . $result->getMessage() . " (DB Error: " . $result->getUserInfo() . ")");
+			}
+			$this->rollbackSQL[] =	"	DELETE FROM	resource_type
+									WHERE		const_name IN ('RESOURCE_TYPE_FILE_RESOURCE_SFTP');";
+		}
+		
+		//	2:	Add the iSeek Data File Resource Type
+		$strSQL	= "	SELECT	id
+					FROM	resource_type
+					WHERE	const_name = 'RESOURCE_TYPE_FILE_IMPORT_CDR_ISEEK_DATA'
+					LIMIT	1;";
+		$result	= $dbAdmin->query($strSQL);
+		if (PEAR::isError($result))
+		{
+			throw new Exception(__CLASS__ . ' Failed to check whether iSeek Data File Resource Type already exists . ' . $result->getMessage() . " (DB Error: " . $result->getUserInfo() . ")");
+		}
+		elseif (!$result->fetchRow())
+		{
+			// DEBUG
+			throw new Exception("SFTP already exists!");
+			// DEBUG
+			
+			$strSQL = "	INSERT INTO	resource_type
+							(name, description, const_name, resource_type_nature)
+						VALUES
+							('iSeek Data Usage File'	, 'iSeek Data Usage File'	, 'RESOURCE_TYPE_FILE_IMPORT_CDR_ISEEK_DATA'	, (SELECT id FROM resource_type_nature WHERE const_name = 'RESOURCE_TYPE_NATURE_IMPORT_FILE' LIMIT 1));";
+			$result = $dbAdmin->query($strSQL);
+			if (PEAR::isError($result))
+			{
+				throw new Exception(__CLASS__ . ' Failed to add the iSeek Data File Resource Type. ' . $result->getMessage() . " (DB Error: " . $result->getUserInfo() . ")");
+			}
+			$this->rollbackSQL[] =	"	DELETE FROM	resource_type
+									WHERE		const_name IN ('RESOURCE_TYPE_FILE_IMPORT_CDR_ISEEK_DATA');";
+		}
 	}
 
 	function rollback()
